requirements.txt
fastapi
uvicorn
pandas
numpy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
requests
